# api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Tetra API",
    "version": "1.0.0",
    "description": "HTTP API for creating, inspecting, administering, and resolving Tetra shortlinks."
  },
  "servers": [
    {
      "url": "/",
      "description": "The server hosting the application"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/users/api-key-info": {
      "get": {
        "operationId": "getAPIKeyInfo",
        "summary": "Get API key regeneration availability",
        "tags": ["Users"],
        "parameters": [
          {"$ref": "#/components/parameters/ApiKeyQuery"},
          {"$ref": "#/components/parameters/ApiKeyHeader"}
        ],
        "security": [
          {"ApiKeyQuery": []},
          {"ApiKeyHeader": []}
        ],
        "responses": {
          "200": {
            "description": "API key regeneration information",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/APIKeyInfo"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/users/regenerate-api-key": {
      "post": {
        "operationId": "regenerateAPIKey",
        "summary": "Regenerate the current user's API key",
        "tags": ["Users"],
        "parameters": [
          {"$ref": "#/components/parameters/ApiKeyQuery"},
          {"$ref": "#/components/parameters/ApiKeyHeader"}
        ],
        "security": [
          {"ApiKeyQuery": []},
          {"ApiKeyHeader": []}
        ],
        "responses": {
          "200": {
            "description": "The newly generated API key",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/RegenerateAPIKeyResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apiKey",
        "description": "User API key. The X-Api-Key header is preferred for non-browser clients."
      },
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "User API key. Used when the apiKey query parameter is absent."
      }
    },
    "parameters": {
      "ApiKeyQuery": {
        "name": "apiKey",
        "in": "query",
        "required": false,
        "description": "API key query parameter. On protected operations, either this parameter or X-Api-Key is required.",
        "schema": {"type": "string", "format": "uuid"}
      },
      "ApiKeyHeader": {
        "name": "X-Api-Key",
        "in": "header",
        "required": false,
        "description": "API key header. Used when apiKey is absent from the query string.",
        "schema": {"type": "string", "format": "uuid"}
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["requestId", "code", "message"],
        "properties": {
          "requestId": {"type": "string", "description": "Request correlation identifier"},
          "code": {"type": "integer", "minimum": 400, "maximum": 599},
          "message": {"type": "string"}
        }
      },
...
      "APIKeyInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": ["canRegenerate", "nextApiKeyAvailable"],
        "properties": {
          "canRegenerate": {"type": "boolean"},
          "nextApiKeyAvailable": {"type": "string", "format": "date-time"}
        }
      },
      "RegenerateAPIKeyResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["apiKey"],
        "properties": {
          "apiKey": {"type": "string", "format": "uuid"}
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication is required or the supplied credentials are invalid",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated user is not permitted to perform this operation",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
...
      "InternalServerError": {
        "description": "An unexpected server error occurred",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      }
    }
  }
}